requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
psycopg2-binary
redis
sqlalchemy
ruff
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
alembic
pytest
httpx
pytest-asyncio